conversion: condition.js to tsx
